+++ Revo.Core/Security/Permission.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using Revo.Core.ValueObjects;
 
@@ -7,19 +8,26 @@ namespace Revo.Core.Security
     {
         public Permission(PermissionType permissionType, string resourceId, string contextId)
         {
-            PermissionType = permissionType;
+            PermissionTypeId = permissionType.Id;
             ResourceId = resourceId;
             ContextId = contextId;
         }
 
+        public Permission(Guid permissionTypeId, string resourceId, string contextId)
+        {
+            PermissionTypeId = permissionTypeId;
+            ResourceId = resourceId;
+            ContextId = contextId;
+        }
+
+        public Guid PermissionTypeId { get; }
         public string ContextId { get; }
-        public PermissionType PermissionType { get; }
         public string ResourceId { get; }
 
         protected override IEnumerable<(string Name, object Value)> GetValueComponents()
         {
             yield return (nameof(ContextId), ContextId);
-            yield return (nameof(PermissionType), PermissionType);
+            yield return (nameof(PermissionTypeId), PermissionTypeId);
             yield return (nameof(ResourceId), ResourceId);
         }
     }