+++ Tests/Revo.Core.Tests/Security/UserPermissionAuthorizerTests.cs
@@ -17,7 +17,7 @@ namespace Revo.Core.Tests.Security
         private FakeUserContext userContext;
         private IPermissionAuthorizationMatcher permissionAuthorizationMatcher;
         private IUserPermissionResolver userPermissionResolver;
-        private IPermissionTypeRegistry permissionTypeRegistry;
+        private Guid permissionType1Id;
         private UserPermissionAuthorizer sut;
 
         public UserPermissionAuthorizerTests()
@@ -25,13 +25,11 @@ namespace Revo.Core.Tests.Security
             userContext = new FakeUserContext();
             permissionAuthorizationMatcher = Substitute.For<IPermissionAuthorizationMatcher>();
             userPermissionResolver = Substitute.For<IUserPermissionResolver>();
-            permissionTypeRegistry = Substitute.For<IPermissionTypeRegistry>();
 
-            permissionTypeRegistry.GetPermissionTypeById(Guid.Parse("0F292EFD-792E-48EC-93DF-CD99EEDB5885"))
-                .Returns(new PermissionType(Guid.Parse("0F292EFD-792E-48EC-93DF-CD99EEDB5885"), "permission"));
+            permissionType1Id = Guid.Parse("0F292EFD-792E-48EC-93DF-CD99EEDB5885");
 
             sut = new UserPermissionAuthorizer(userContext, permissionAuthorizationMatcher,
-                userPermissionResolver, permissionTypeRegistry);
+                userPermissionResolver);
         }
 
         [Theory]
@@ -52,8 +50,7 @@ namespace Revo.Core.Tests.Security
                 {
                     var requiredPermissions = ci.ArgAt<IReadOnlyCollection<Permission>>(1);
                     if (requiredPermissions.Count == 1 &&
-                        requiredPermissions.First().PermissionType.Equals(permissionTypeRegistry.GetPermissionTypeById(
-                            Guid.Parse("0F292EFD-792E-48EC-93DF-CD99EEDB5885")))
+                        requiredPermissions.First().PermissionTypeId == permissionType1Id
                         && requiredPermissions.First().ContextId == "context_id" &&
                         requiredPermissions.First().ResourceId == "resource_id")
                     {
@@ -63,7 +60,7 @@ namespace Revo.Core.Tests.Security
                     throw new ArgumentException();
                 });
 
-            bool result = await sut.CheckAuthorizationAsync(user, "0F292EFD-792E-48EC-93DF-CD99EEDB5885",
+            bool result = await sut.CheckAuthorizationAsync(user, Guid.Parse("0F292EFD-792E-48EC-93DF-CD99EEDB5885"),
                 "resource_id", "context_id");
             result.Should().Be(passes);
         }
@@ -83,8 +80,7 @@ namespace Revo.Core.Tests.Security
                 {
                     var requiredPermissions = ci.ArgAt<IReadOnlyCollection<Permission>>(1);
                     if (requiredPermissions.Count == 1 &&
-                        requiredPermissions.First().PermissionType.Equals(permissionTypeRegistry.GetPermissionTypeById(
-                            Guid.Parse("0F292EFD-792E-48EC-93DF-CD99EEDB5885")))
+                        requiredPermissions.First().PermissionTypeId == permissionType1Id
                         && requiredPermissions.First().ContextId == "context_id" &&
                         requiredPermissions.First().ResourceId == "resource_id")
                     {
@@ -94,7 +90,7 @@ namespace Revo.Core.Tests.Security
                     throw new ArgumentException();
                 });
 
-            bool result = await sut.CheckCurrentUserAuthorizationAsync("0F292EFD-792E-48EC-93DF-CD99EEDB5885",
+            bool result = await sut.CheckCurrentUserAuthorizationAsync(Guid.Parse("0F292EFD-792E-48EC-93DF-CD99EEDB5885"),
                 "resource_id", "context_id");
             result.Should().Be(passes);
         }